# assets/json/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "services-registry",
    "description": "Sails.js based services registry to decouple front end applications from their backend services",
    "version": "1.0.0"
  },
  "host": "localhost:1337",
  "tags": [
    {
      "name": "applications",
      "description": "Manage applications"
    },
    {
      "name": "configs",
      "description": "Manage applications configurations"
    }
  ],
  "paths": {
    "/applications": {
      "get": {
        "summary": "Get all registered applications with their configurations",
        "description": "Get all registered applications with their configurations",
        "operationId": "getApplications",
        "tags": [
          "applications"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "A list of applications",
            "schema": {
              "type": "array"
            }
          }
        }
      },
      "post": {
        "summary": "Add an application with its optional configurations",
        "description": "Add an application with its optional configurations",
        "operationId": "addApplication",
        "tags": [
          "applications"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Application that needs to be added",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Payload for application creation",
            "schema": {
              "type": "object"
            }
          }
        }
      },
      "put": {
        "summary": "Update an existing application",
        "description": "Update an existing application",
        "operationId": "updateApplication",
        "tags": [
          "applications"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Payload for application update",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The application that has been just added",
            "schema": {
              "type": "object"
            }
          },
          "404": {
            "description": "The application could not be found"
          }
        }
      }
    },
    "/applications/{app}": {
      "get": {
        "summary": "Get all configurations for a given application identified by its name",
        "description": "Get all configurations for a given application identified by its name",
        "operationId": "getApplication",
        "tags": [
          "applications"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "app",
            "in": "path",
            "description": "Name of the application to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The application with its configurations is found"
          },
          "404": {
            "description": "The application could not be found"
          }
        }
      },
      "delete": {
        "summary": "Delete an existing application identified by its name",
        "description": "Delete an existing application identified by its name",
        "operationId": "removeApplication",
        "tags": [
          "applications"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "app",
            "in": "path",
            "description": "Name of the application to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The application has been successfully deleted"
          },
          "404": {
            "description": "The application could not be found"
          }
        }
      }
    },
    "/applications/{app}/configs": {
      "get": {
        "summary": "Get all configurations for a given application identified by its name",
        "description": "Get all configurations for a given application identified by its name",
        "operationId": "getConfigs",
        "tags": [
          "configs"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "app",
            "in": "path",
            "description": "Name of the application to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The application with its configurations is found"
          },
          "404": {
            "description": "The application could not be found"
          }
        }
      },
      "put": {
        "summary": "Update application configurations",
        "description": "Update application configurations",
        "operationId": "setConfigs",
        "tags": [
          "configs"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "app",
            "in": "path",
            "description": "Name of the concerned application",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Payload for configurations update",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The application configurations have been successfully updated"
          },
          "404": {
            "description": "The application could not be found"
          }
        }
      },
      "delete": {
        "summary": "Delete all application configurations",
        "description": "Delete all application configurations",
        "operationId": "removeConfigs",
        "tags": [
          "configs"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "app",
            "in": "path",
            "description": "Name of the concerned application",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The application configurations have been successfully deleted"
          },
          "404": {
            "description": "The application could not be found"
          }
        }
      }
    }
  }
}